remove unused collect_planet function

+++ lib/contants/constants.h
@@ -15,17 +15,17 @@
 #define MAX_SPEED_CHEAT 10000
 #define MAX_COLLECTOR_SPEED 1000
 
-#define ATTACKER_1 1
-#define ATTACKER_2 2
-#define ATTACKER_3 3
-#define ATTACKER_4 4
-#define ATTACKER_5 5
+#define ATTACKER_1 (uint8_t)1
+#define ATTACKER_2 (uint8_t)2
+#define ATTACKER_3 (uint8_t)3
+#define ATTACKER_4 (uint8_t)4
+#define ATTACKER_5 (uint8_t)5
 
-#define EXPLORER_1 6
-#define EXPLORER_2 7
+#define EXPLORER_1 (uint8_t)6
+#define EXPLORER_2 (uint8_t)7
 
-#define COLLECTOR_1 8
-#define COLLECTOR_2 9
+#define COLLECTOR_1 (uint8_t)8
+#define COLLECTOR_2 (uint8_t)9
 
 #define MAX_PLANETS_NUMBER 8
 #define SHIPS_NUMBER 9
